moved duplicate code in tools module
