+++ tests/framework/model/test_multi_gpu.py
@@ -5,65 +5,11 @@ from unittest import skipIf
 import torch
 from torch import nn
 
-from poutyne import Model, EpochMetric
+from poutyne import Model
 from .base import ModelFittingTestCase
-
-some_metric_1_value = 1.
-some_metric_2_value = 2.
-repeat_batch_metric_value = 3.
-
-
-def some_batch_metric_1(y_pred, y_true):
-    return torch.FloatTensor([some_metric_1_value])
-
-
-def some_batch_metric_2(y_pred, y_true):
-    return torch.FloatTensor([some_metric_2_value])
-
-
-def repeat_batch_metric(y_pred, y_true):
-    return torch.FloatTensor([repeat_batch_metric_value])
-
-
-class SomeEpochMetric(EpochMetric):
-
-    def __init__(self):
-        super().__init__()
-        self.increment = 0.0
-
-    def forward(self, y_pred, y_true):
-        self.increment += 1
-
-    def get_metric(self):
-        increment_value = self.increment
-        self.increment = 0
-        return increment_value
-
-    def reset(self):
-        pass
-
-
-some_constant_epoch_metric_value = 3
-
-
-class SomeConstantEpochMetric(EpochMetric):
-
-    def forward(self, y_pred, y_true):
-        pass
-
-    def get_metric(self):
-        return torch.FloatTensor([some_constant_epoch_metric_value])
-
-    def reset(self):
-        pass
-
-
-def some_data_tensor_generator(batch_size):
-    while True:
-        x = torch.rand(batch_size, 1)
-        y = torch.rand(batch_size, 1)
-        yield x, y
-
+from ..tools import some_data_tensor_generator, some_batch_metric_1, some_batch_metric_2, repeat_batch_metric, \
+    some_metric_1_value, some_metric_2_value, repeat_batch_metric_value, SomeConstantEpochMetric, \
+    some_constant_epoch_metric_value
 
 TEST_MULTI_GPUS = int(os.environ.get('MULTI_GPUS', 0))
