+++ tests/framework/tools.py
@@ -1,0 +1,60 @@
+# pylint: disable=unused-argument
+
+import torch
+
+from poutyne import EpochMetric
+
+some_metric_1_value = 1.
+some_metric_2_value = 2.
+repeat_batch_metric_value = 3.
+
+some_constant_epoch_metric_value = 3
+
+
+def some_data_tensor_generator(batch_size):
+    while True:
+        x = torch.rand(batch_size, 1)
+        y = torch.rand(batch_size, 1)
+        yield x, y
+
+
+def some_batch_metric_1(y_pred, y_true):
+    return torch.FloatTensor([some_metric_1_value])
+
+
+def some_batch_metric_2(y_pred, y_true):
+    return torch.FloatTensor([some_metric_2_value])
+
+
+def repeat_batch_metric(y_pred, y_true):
+    return torch.FloatTensor([repeat_batch_metric_value])
+
+
+class SomeConstantEpochMetric(EpochMetric):
+
+    def forward(self, y_pred, y_true):
+        pass
+
+    def get_metric(self):
+        return torch.FloatTensor([some_constant_epoch_metric_value])
+
+    def reset(self):
+        pass
+
+
+class SomeEpochMetric(EpochMetric):
+
+    def __init__(self):
+        super().__init__()
+        self.increment = 0.0
+
+    def forward(self, y_pred, y_true):
+        self.increment += 1
+
+    def get_metric(self):
+        increment_value = self.increment
+        self.increment = 0
+        return increment_value
+
+    def reset(self):
+        pass
